+++ wgpu-raw-hs/src/WGPU/Raw/GLFWSurface.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE ForeignFunctionInterface #-}
+{-# OPTIONS_GHC -Wno-unused-imports #-}
 
 module WGPU.Raw.GLFWSurface where
 
@@ -9,10 +10,11 @@ import qualified WGPU.Raw.Generated.Enum.WGPUSType as WGPUSType
 import WGPU.Raw.Generated.Fun (WGPUHsInstance, wgpuInstanceCreateSurface)
 import WGPU.Raw.Generated.Struct.WGPUChainedStruct
 import WGPU.Raw.Generated.Struct.WGPUSurfaceDescriptor
-import WGPU.Raw.Generated.Struct.WGPUSurfaceDescriptorFromMetalLayer
 import WGPU.Raw.Types (WGPUInstance (WGPUInstance), WGPUSurface)
 
 #ifdef WGPUHS_TARGET_MACOS
+
+import WGPU.Raw.Generated.Struct.WGPUSurfaceDescriptorFromMetalLayer
 
 createSurface ::
   WGPUHsInstance ->
@@ -70,5 +72,57 @@ createSurface ::
   GLFW.Window ->
   IO WGPUSurface
 createSurface inst window = error "Linux: not yet implemented."
+
+#endif
+
+#ifdef WGPUHS_TARGET_WINDOWS
+
+import System.Win32.DLL (getModuleHandle)
+import WGPU.Raw.Generated.Struct.WGPUSurfaceDescriptorFromWindowsHWND
+
+createSurface ::
+  WGPUHsInstance ->
+  GLFW.Window ->
+  IO WGPUSurface
+createSurface inst window = do
+  hWnd <- GLFW.getWin32Window window
+  hInstance <- getModuleHandle Nothing
+
+  alloca $ \ptr_surfaceDescriptor -> do
+    alloca $ \ptr_chainedStruct -> do
+      alloca $ \ptr_surfaceDescriptorFromWindowHWND -> do
+
+        let surfaceDescriptorFromWindowHWND =
+              WGPUSurfaceDescriptorFromWindowsHWND
+              { chain =
+                  WGPUChainedStruct
+                  { next = nullPtr,
+                    sType = WGPUSType.SurfaceDescriptorFromWindowsHWND
+                  },
+                hinstance = hInstance,
+                hwnd = hWnd
+              }
+        poke
+          ptr_surfaceDescriptorFromWindowHWND
+          surfaceDescriptorFromWindowHWND
+
+        let chainedStruct =
+             WGPUChainedStruct
+               { next = castPtr ptr_surfaceDescriptorFromWindowHWND,
+                 sType = WGPUSType.SurfaceDescriptorFromWindowsHWND
+               }
+        poke ptr_chainedStruct chainedStruct
+
+        let surfaceDescriptor =
+              WGPUSurfaceDescriptor
+                { nextInChain = ptr_chainedStruct,
+                  label = nullPtr
+                }
+        poke ptr_surfaceDescriptor surfaceDescriptor
+
+        wgpuInstanceCreateSurface
+          inst
+          (WGPUInstance nullPtr)
+          ptr_surfaceDescriptor
 
 #endif