+++ wgpu-hs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for wgpu-hs
 
+## 0.2.0.0 -- 2021-08-22
+
+- Added MS Windows support for surfaces and DLL loading.
+- Added `withPlatformInstance`: select dynamic library name based on platform.
+
 ## 0.1.0.0 -- 2021-08-20
 
 - Initial (incomplete) API bindings started.